+++ xcube/core/store/fs/registry.py
@@ -28,11 +28,12 @@ from .accessor import FsDataAccessor
 from .impl.dataset import DatasetGeoTiffFsDataAccessor
 from .impl.dataset import DatasetNetcdfFsDataAccessor
 from .impl.dataset import DatasetZarrFsDataAccessor
+from .impl.fs import AzureFsAccessor
 from .impl.fs import FileFsAccessor
 from .impl.fs import FtpFsAccessor
 from .impl.fs import MemoryFsAccessor
+from .impl.fs import ReferenceFsAccessor
 from .impl.fs import S3FsAccessor
-from .impl.fs import AzureFsAccessor
 from .impl.geodataframe import GeoDataFrameGeoJsonFsDataAccessor
 from .impl.geodataframe import GeoDataFrameShapefileFsDataAccessor
 from .impl.geotiff import MultiLevelDatasetGeoTiffFsDataAccessor
@@ -62,8 +63,12 @@ def register_fs_accessor_class(
 
 
 for cls in (
-        FileFsAccessor, S3FsAccessor, AzureFsAccessor, MemoryFsAccessor,
-        FtpFsAccessor
+        AzureFsAccessor,
+        FileFsAccessor,
+        FtpFsAccessor,
+        MemoryFsAccessor,
+        ReferenceFsAccessor,
+        S3FsAccessor,
 ):
     register_fs_accessor_class(cls)